keeper: support --wal-dir
For those who run Postgres with a WAL directory symlinked to a separate
disk (often for performance reasons) this commit adds a --wal-dir flag
to the keeper.

The symlink to a new WAL disk needs to be created as part of the
database initialisation step, and also whenever we clobber the database,
such as when taking a pg_basebackup (if pg_rewind is
disabled/unavailable). By passing this directory as an argument to
initdb and pg_basebackup, we can cover all the scenarios that the keeper
erases and recreates the data directory, ensuring the symlink is always
present.