Support waldir in pitr command
For those who use different devices for wal and data directory, we
should support the interpolation of the wal directory location in the
pitr command as we do the data directory. This allows people to supply
commands like:

```
pg_basebackup --pgdata %d --waldir %w ...
```

And have both data directory and wal directory templated. pg_basebackup
and initdb are commands that support the concept of waldir, but this
could be useful for any arbitrary pitr command.

The commit also simplifies the implementation of the expand function to
feature less mutation and slice offset accesses.