Force checkpoint before rewind
Prior to running pg_rewind, force a checkpoint on the source database to
ensure the pg_control file used by pg_rewind is up-to-date. Failure to
do so can result in pg_rewind not detecting a timeline fork and exiting
with no action, leading stolon to fallback on pg_basebackup (which
itself will checkpoint) to ensure a resync is successful.