Cascading replication against sync standbys
When synchronous standbys are enabled, it makes sense to cascade
replication from the sync rather than the primary. Not only does this
reduce the load on the primary- which is important, as replicating to
many replicas can cause a spike in network traffic which impacts sync
acknowledgement latency- but it increases the likelihood of timeline
forks as the async replicas can run ahead of the sync.

This commit changes the sentinel to have standbys follow from
synchronous standbys by default, falling back to the master if none are
available. We'll need to parameterise this and think about more generic
ways of specifying cascading rules, but this is a good start.

```
===== Keepers/DB tree =====

stolon_production_2 (master)
└─stolon_production_1 (sync)
  └─stolon_production_0 (async)
```