Add pg_rewind interval and timeout to clusterspec
Allow the setting of the a pg_rewind interval and timeout so that the
retry behaviour of pg_rewind can be customised. The defualt values if
these are not set in the clusterspec are 0s which results in the
current default behaviour of only trying to pg_rewind once if
usePgrewind is set to true.