Hardcode pg_rewind retrying (up-to 5m)
1. https://github.com/gocardless/stolon/pull/14

Retry pg_rewind for up-to 5m after the first attempt. This allows us to
handle the followed database booting for a moment before falling back to
the more expensive pg_basebackup.

This comes as a crappier alternative to [1], where we were creating
proper configuration values. We want to try this before we attempt to
create a more mature setup, as we intend to leave that work for whenever
we upstream our cascading replication.