Add new keeper flags
--never-master, prevent keeper from being elected as master
--never-synchronous-replica, prevent keeper from being chosen as
synchronous replica

We believe that it makes more sense to have different flags, to keep
consistency with the rest of the code.

There is a `KeeperSpec` that is empty which we try to use, but it
required to add these two new flags to the `KeeperInfo` that is part of
the cluster state. Then the sentinel can make decisions based on these
new flags