Allow keeper to recover via archive recovery
When restoring a cluster from a set of disk snapshots, that all
originate from the same machine, we want to ensure that Stolon brings up
the cluster without doing a full 'resync', either via `pg_rewind` or
`pg_basebackup`.

The reasons for this are:
1. `pg_rewind` will refuse to run if the target (the instance which is
being rewound) has already replayed all of the WAL that it had received.
2. `pg_basebackup` is very expensive in terms of time and bandwidth,
when there is a large volume of data.