Add option to keep data in pitr init_mode
This option is useful if the there is already a valid PGDATA directory
on disk, in the location that the Keeper is managing, and you would like
to keep this when performing a Point In Time Recovery.

We will use this for the case of a new Stolon cluster where the keeper
instances have disks that have been created from a snapshot of a keeper
disk in a different cluster.