Unset recovery parameters after restoring in PITR init mode
The recovery params being set will write out a ``recovery.conf`` any time the
Postgres database is reloaded, restarted or started in any way. This
will be detected by stolon as the node being on standby. This causes a
promotion to be attempted, which fails as postgres is already running
as master. Due to this, we end up in a loop where the node is
attempted to be promoted, and failures cause it to try again.